+++ tests/dummy/app/shadowed-file.ts
@@ -1,5 +1,0 @@
-// This file should win out over `shadowed-file.ts` and `shadowed-file.js`
-// in our two in-repo addons.
-const value: string = 'dummy/shadowed-file';
-
-export default value;