+++ tests/dummy/lib/in-repo-a/app/shadowed-file.ts
@@ -1,4 +1,0 @@
-// This should be shadowed by the dummy app's own `shadowed-file.ts`
-const value: string = 'bad';
-
-export default value;