+++ tests/dummy/lib/in-repo-b/app/shadowed-file.js
@@ -1,4 +1,0 @@
-// This should be shadowed by the dummy app's own `sahdowed-file.ts`
-const value = 'bad';
-
-export default value;