+++ ts/addon.ts
@@ -1,8 +1,8 @@
 import semver from 'semver';
 import { Remote } from 'stagehand';
 import { connect } from 'stagehand/lib/adapters/child-process';
-import { hasPlugin, addPlugin, AddPluginOptions } from 'ember-cli-babel-plugin-helpers';
 import Addon from 'ember-cli/lib/models/addon';
+import PreprocessRegistry from 'ember-cli-preprocess-registry';
 import { addon } from './lib/utilities/ember-cli-entities';
 import fork from './lib/utilities/fork';
 import TypecheckWorker from './lib/typechecking/worker';
@@ -10,6 +10,9 @@ import TypecheckMiddleware from './lib/typechecking/middleware';
 import { Application } from 'express';
 import walkSync from 'walk-sync';
 import fs from 'fs-extra';
+import logger from 'debug';
+
+const debug = logger('ember-cli-typescript:addon');
 
 export const ADDON_NAME = 'ember-cli-typescript';
 
@@ -18,6 +21,7 @@ export default addon({
 
   included() {
     this._super.included.apply(this, arguments);
+
     this._checkDevelopment();
     this._checkAddonAppFiles();
     this._checkBabelVersion();
@@ -44,12 +48,22 @@ export default addon({
     return `${__dirname}/blueprints`;
   },
 
-  serverMiddleware({ app }) {
-    this._addTypecheckMiddleware(app);
+  serverMiddleware({ app, options }) {
+    if (!options || !options.path) {
+      debug('Installing typecheck server middleware');
+      this._addTypecheckMiddleware(app);
+    } else {
+      debug('Skipping typecheck server middleware');
+    }
   },
 
-  testemMiddleware(app) {
-    this._addTypecheckMiddleware(app);
+  testemMiddleware(app, options) {
+    if (!options || !options.path) {
+      debug('Installing typecheck testem middleware');
+      this._addTypecheckMiddleware(app);
+    } else {
+      debug('Skipping typecheck testem middleware');
+    }
   },
 
   async postBuild() {
@@ -69,26 +83,15 @@ export default addon({
     }
   },
 
-  setupPreprocessorRegistry(type) {
+  setupPreprocessorRegistry(type, registry) {
     if (type !== 'parent') return;
 
-    // Normally this is the sort of logic that would live in `included()`, but
-    // ember-cli-babel reads the configured extensions when setting up the
-    // preprocessor registry, so we need to beat it to the punch.
-    this._registerBabelExtension();
-
-    // As of 3.7, TS supports the optional chaining and nullish coalescing proposals.
-    // https://devblogs.microsoft.com/typescript/announcing-typescript-3-7-beta/
-    // Since we can't necessarily know what version of TS an addon was developed with,
-    // we unconditionally add the Babel plugins for both proposals.
-    this._addBabelPluginIfNotPresent('@babel/plugin-proposal-optional-chaining');
-    this._addBabelPluginIfNotPresent('@babel/plugin-proposal-nullish-coalescing-operator');
-
-    // Needs to come after the class properties plugin (see tests/unit/build-test.ts -
-    // "property initialization occurs in the right order")
-    this._addBabelPluginIfNotPresent('@babel/plugin-transform-typescript', {
-      after: ['@babel/plugin-proposal-class-properties'],
-    });
+    // If we're acting on behalf of the root app, issue a warning if we detect
+    // a situation where a .js file from an addon has the same name as a .ts
+    // file in the app, as which file wins is nondeterministic.
+    if (this.parent === this.project) {
+      this._registerCollisionDetectionPreprocessor(registry);
+    }
   },
 
   shouldIncludeChildAddon(addon) {
@@ -97,13 +100,55 @@ export default addon({
     return !['in-repo-a', 'in-repo-b'].includes(addon.name);
   },
 
+  _registerCollisionDetectionPreprocessor(registry: PreprocessRegistry) {
+    registry.add('js', {
+      name: 'ember-cli-typescript-collision-check',
+      toTree: (input, path) => {
+        if (path !== '/') return input;
+
+        let addon = this;
+        let checked = false;
+        let stew = require('broccoli-stew') as typeof import('broccoli-stew');
+
+        return stew.afterBuild(input, function () {
+          if (!checked) {
+            checked = true;
+            addon._checkForFileCollisions(this.inputPaths[0]);
+          }
+        });
+      },
+    });
+  },
+
+  _checkForFileCollisions(directory: string) {
+    let walkSync = require('walk-sync') as typeof import('walk-sync');
+    let files = new Set(walkSync(directory, ['**/*.{js,ts}']));
+
+    let collisions = [];
+    for (let file of files) {
+      if (file.endsWith('.js') && files.has(file.replace(/\.js$/, '.ts'))) {
+        collisions.push(file.replace(/\.js$/, '.{js,ts}'));
+      }
+    }
+
+    if (collisions.length) {
+      this.ui.writeWarnLine(
+        'Detected collisions between .js and .ts files of the same name. ' +
+          'This can result in nondeterministic build output; ' +
+          'see https://git.io/JvIwo for more information.\n  - ' +
+          collisions.join('\n  - ')
+      );
+    }
+  },
+
   _checkBabelVersion() {
-    let babel = this.parent.addons.find(addon => addon.name === 'ember-cli-babel');
+    let babel = this.parent.addons.find((addon) => addon.name === 'ember-cli-babel');
     let version = babel && babel.pkg.version;
-    if (!babel || !(semver.gte(version!, '7.7.3') && semver.lt(version!, '8.0.0'))) {
+
+    if (!babel || !(semver.gte(version!, '7.17.0') && semver.lt(version!, '8.0.0'))) {
       let versionString = babel ? `version ${babel.pkg.version}` : `no instance of ember-cli-babel`;
       this.ui.writeWarnLine(
-        `ember-cli-typescript requires ember-cli-babel ^7.7.3, but you have ${versionString} installed; ` +
+        `ember-cli-typescript requires ember-cli-babel ^7.17.0, but you have ${versionString} installed; ` +
           'your TypeScript files may not be transpiled correctly.'
       );
     }
@@ -164,31 +209,6 @@ export default addon({
     }
   },
 
-  _getConfigurationTarget() {
-    // If `this.app` isn't present, we know `this.parent` is an addon
-    return this.app || (this.parent as Addon);
-  },
-
-  _registerBabelExtension() {
-    let target = this._getConfigurationTarget();
-    let options: Record<string, any> = target.options || (target.options = {});
-    let babelAddonOptions: Record<string, any> =
-      options['ember-cli-babel'] || (options['ember-cli-babel'] = {});
-    let extensions: string[] =
-      babelAddonOptions.extensions || (babelAddonOptions.extensions = ['js']);
-
-    if (!extensions.includes('ts')) {
-      extensions.push('ts');
-    }
-  },
-
-  _addBabelPluginIfNotPresent(pluginName: string, pluginOptions?: AddPluginOptions) {
-    let target = this._getConfigurationTarget();
-    if (!hasPlugin(target, pluginName)) {
-      addPlugin(target, require.resolve(pluginName), pluginOptions);
-    }
-  },
-
   _addTypecheckMiddleware(app: Application) {
     let workerPromise = this._getTypecheckWorker();
     let middleware = new TypecheckMiddleware(this.project, workerPromise);
@@ -209,7 +229,7 @@ export default addon({
     let childProcess = fork(`${__dirname}/lib/typechecking/worker/launch`);
     let worker = await connect<TypecheckWorker>(childProcess);
 
-    await worker.onTypecheck(status => {
+    await worker.onTypecheck((status) => {
       for (let error of status.errors) {
         this.ui.writeLine(error);
       }